+++ parquet-hadoop/src/main/java/org/apache/parquet/hadoop/util/H2SeekableInputStream.java
@@ -25,7 +25,7 @@ import java.nio.ByteBuffer;
 import java.util.List;
 import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.parquet.bytes.ByteBufferAllocator;
-import org.apache.parquet.hadoop.util.vectorio.VectorIOBridge;
+import org.apache.parquet.hadoop.util.wrappedio.VectorIOBridge;
 import org.apache.parquet.io.DelegatingSeekableInputStream;
 import org.apache.parquet.io.ParquetFileRange;
