+++ parquet-hadoop/src/main/java/org/apache/parquet/hadoop/util/wrappedio/VectorIOBridge.java
@@ -16,12 +16,12 @@
  * limitations under the License.
  */
 
-package org.apache.parquet.hadoop.util.vectorio;
+package org.apache.parquet.hadoop.util.wrappedio;
 
 import static java.util.Objects.requireNonNull;
 import static org.apache.parquet.Exceptions.throwIfInstance;
 import static org.apache.parquet.Preconditions.checkArgument;
-import static org.apache.parquet.hadoop.util.vectorio.BindingUtils.loadInvocation;
+import static org.apache.parquet.hadoop.util.wrappedio.BindingUtils.loadInvocation;
 
 import java.io.EOFException;
 import java.io.IOException;
